Make TNGFile.seek raise the expected error
Fixes #11

The message attached to the exception raised by TNGFile.seek when a file
is opened in write mode is now consistent with what happens. The same
goes for the message for the exception raised when seeking on a non-open
file.

The test for the message about write mode is skipped at the moment
because of #10.