First try at TNG writing
Adds a `write` method to the `TNGFile` class. At that point, this method
successfully writes the positions and the box shape. It tries to write
the time, but fails. It does not try to write velocities and forces yet,
nor does it ties to write the topology.

The written file can be read by pytng. It is also read by `gmx check`,
but yields the following message:

    TNG library: Cannot uncompress data block. /home/jon/src/gromacs-2016.3/src/external/tng_io/src/lib/tng_io.c:5298